Big 3 Programmers said everything look good, but sim wont pop up motor
On everything I love I will see this work, even if it kills me. In this life that I live I will not settle for half-assed programs that I do not see results from. I did not spend more than 10 fucking hours on this bs to not see it work.